requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
requests==2.32.3
pandas==2.2.2
numpy==1.26.4
aiofiles==23.2.1
